requirements.txt
ccxt
pyrebase
python-dotenv
websocket-client
pycryptodome
requests
